requirements.txt
nicegui
fastapi
uvicorn[standard]
python-multipart
supabase
pydantic-settings
openai
python-docx
python-dotenv
PyPDF2
pdfplumber
pymupdf
fpdf2
